Setup TYPO3 as user `www-data` instead of `root`
https://getcomposer.org/doc/faqs/how-to-install-untrusted-packages-safely.md

"Certain Composer commands, including exec, install, and update allow third party code to execute on your system. This is from its "plugins" and "scripts" features. Plugins and scripts have full access to the user account which runs Composer. For this reason, it is strongly advised to avoid running Composer as super-user/root."

Signed-off-by: Christos Sidiropoulos <csidirop@runbox.com>